+++ CHANGELOG.md
@@ -10,5 +10,17 @@
 ### Changes
 - Fixed an error that would sometimes happen on rfupf window leading it to close
 - Made the announcement notification trigger upon joining SkyBlock instead of upon joining the server
+- Adjusted the daily challenges weights:
+  - Party Fisher: 100 → 35
+  - Sea Creature Hunter: 50 → 75
+  - Coccon Master: 25 → 50
+  - Precious Trophy: 100 → 25
+  - Hotspot Hunter: 50 → 75
+  - All Island Tasks 10 → 20
+    - Exceptions:
+      - Desert 10 → 5
+      - Hub 10 → 5
+      - Spider 10 → 5
+      - Dwarven 10 → 5
 
 ### Back-end